api/requirements.txt
Flask==3.1.1
Flask_Cors==5.0.0
mariadb==1.1.13
Pillow==11.3.0
python-dotenv==1.1.1
Requests==2.32.4
